Feature/#152: 컴포넌트 분리
